refactor(cart): extract CartSaveLoadCoordinator
Move Save / Save-as / Load / Rename / Delete flows plus the
unsaved-changes prompt into their own coordinator. CartActivity keeps
only the wiring — a menu tap or on-back callback now delegates to the
coordinator instead of hosting the dialog logic inline.
